refactor: :recycle: перенести обработку внешних файлов
Общая логика externalFile относится к property orchestration и не должна зависеть от DynamicList. Перенос сохраняет поведение XML/YAML и убирает обратную зависимость от реализации формы.
